AIBot/requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
watchdog==3.0.0
PyMuPDF==1.21.1
pandas==1.5.3
Unidecode==1.3.6
torch==2.0.0
sentence-transformers==2.2.2
openai==0.27.2
spacy==3.5.1
langdetect==1.0.9
iso-639==0.4.5
keybert==0.7.0
pdfminer.six==20221105
sortedcontainers==2.2.2
transformers==4.28.1
faiss-cpu==1.7.4
openpyxl==3.1.2